+++ gen/aws/drivers_definitions.go
@@ -969,6 +969,10 @@ var DriversDefs = []driversDef{
 		},
 	},
 	{
+		Api:     "acm",
+		Drivers: []driver{},
+	},
+	{
 		Api:     "sts",
 		Drivers: []driver{},
 	},