+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### AWS Services
 
 - Update IAM policies, to add statements with `awless update policy`
+- Add ACM certificates in infra: `awless list certificates` 
 
 ### Fixes
 - `awless ssh --through`: no reusing same conn to avoid EOF. Bug: only first user (amazonlinux) was successful (usually ec2-user) !!